[#51] Rename folder from app/config to app/core [#51] Update npm packages
