+++ app/package.json
@@ -18,13 +18,13 @@
 	"dependencies": {
 		"@emotion/react": "^11.14.0",
 		"@mantine/core": "^8.3.14",
-		"@tanstack/react-query": "^5.90.20",
+		"@tanstack/react-query": "^5.90.21",
 		"@tanstack/react-table": "^8.21.3",
 		"@tanstack/react-virtual": "^3.13.18",
 		"clsx": "^2.1.1",
 		"flag-icons": "^7.5.0",
 		"i18n-iso-countries": "^7.14.0",
-		"i18next": "^25.8.4",
+		"i18next": "^25.8.7",
 		"lodash": "^4.17.23",
 		"next": "16.1.6",
 		"react": "19.2.4",
@@ -35,27 +35,27 @@
 	},
 	"devDependencies": {
 		"@faker-js/faker": "^10.3.0",
-		"@storybook/addon-a11y": "^10.2.7",
-		"@storybook/addon-docs": "^10.2.7",
-		"@storybook/addon-vitest": "^10.2.7",
-		"@storybook/nextjs-vite": "^10.2.7",
-		"@storybook/react": "^10.2.7",
+		"@storybook/addon-a11y": "^10.2.8",
+		"@storybook/addon-docs": "^10.2.8",
+		"@storybook/addon-vitest": "^10.2.8",
+		"@storybook/nextjs-vite": "^10.2.8",
+		"@storybook/react": "^10.2.8",
 		"@svgr/webpack": "^8.1.0",
 		"@tailwindcss/postcss": "^4.1.18",
-		"@types/node": "^25.2.1",
-		"@types/react": "^19.2.13",
+		"@types/node": "^25.2.3",
+		"@types/react": "^19.2.14",
 		"@types/react-dom": "^19.2.3",
 		"@vitest/browser-playwright": "^4.0.18",
 		"eslint": "^9.39.2",
 		"eslint-config-next": "16.1.6",
-		"eslint-plugin-storybook": "^10.2.7",
+		"eslint-plugin-storybook": "^10.2.8",
 		"playwright": "^1.58.2",
 		"prettier": "^3.8.1",
 		"prettier-plugin-css-order": "^2.2.0",
 		"prettier-plugin-organize-imports": "^4.3.0",
 		"prettier-plugin-tailwindcss": "^0.7.2",
 		"remark-gfm": "^4.0.1",
-		"storybook": "^10.2.7",
+		"storybook": "^10.2.8",
 		"tailwindcss": "^4.1.18",
 		"typescript": "^5.9.3",
 		"vite": "^7.3.1",
