+++ app/src/core/i18n/i18n.d.ts
