+++ app/tsconfig-app.json
@@ -12,8 +12,8 @@
 		"paths": {
 			"@app-components": ["./src/shared/components/index.ts"],
 			"@app-utils": ["./src/shared/utils/index.ts"],
-			"@app-i18n": ["./src/config/i18n/index.ts"],
-			"@app-api": ["./src/config/api/index.ts"],
+			"@app-i18n": ["./src/core/i18n/index.ts"],
+			"@app-api": ["./src/core/api/index.ts"],
 			"@app/*": ["./src/*"],
 			"@ds/*": ["../ds/dist/*"]
 		}
