+++ ds/package.json
@@ -21,18 +21,18 @@
 	"devDependencies": {
 		"@eslint/js": "^10.0.1",
 		"@faker-js/faker": "^10.3.0",
-		"@storybook/addon-a11y": "^10.2.7",
-		"@storybook/addon-docs": "^10.2.7",
-		"@storybook/addon-vitest": "^10.2.7",
-		"@storybook/react-vite": "^10.2.7",
+		"@storybook/addon-a11y": "^10.2.8",
+		"@storybook/addon-docs": "^10.2.8",
+		"@storybook/addon-vitest": "^10.2.8",
+		"@storybook/react-vite": "^10.2.8",
 		"@tailwindcss/postcss": "^4.1.18",
 		"@types/eslint": "^9.6.1",
 		"@types/estree": "^1.0.8",
 		"@types/lodash": "^4.17.23",
-		"@types/node": "^25.2.1",
-		"@types/react": "^19.2.13",
+		"@types/node": "^25.2.3",
+		"@types/react": "^19.2.14",
 		"@types/react-dom": "^19.2.3",
-		"@vitejs/plugin-react": "^5.1.3",
+		"@vitejs/plugin-react": "^5.1.4",
 		"@vitest/browser-playwright": "^4.0.18",
 		"@vitest/coverage-v8": "^4.0.18",
 		"eslint": "^10.0.0",
@@ -40,7 +40,7 @@
 		"eslint-plugin-import": "^2.32.0",
 		"eslint-plugin-react-hooks": "^7.0.1",
 		"eslint-plugin-react-refresh": "^0.5.0",
-		"eslint-plugin-storybook": "^10.2.7",
+		"eslint-plugin-storybook": "^10.2.8",
 		"globals": "^17.3.0",
 		"playwright": "^1.58.2",
 		"prettier": "^3.8.1",
@@ -48,10 +48,10 @@
 		"prettier-plugin-organize-imports": "^4.3.0",
 		"prettier-plugin-tailwindcss": "^0.7.2",
 		"remark-gfm": "^4.0.1",
-		"storybook": "^10.2.7",
+		"storybook": "^10.2.8",
 		"tailwindcss": "^4.1.18",
 		"typescript": "~5.9.3",
-		"typescript-eslint": "^8.54.0",
+		"typescript-eslint": "^8.55.0",
 		"vite": "^7.3.1",
 		"vite-plugin-svgr": "^4.5.0",
 		"vitest": "^4.0.18"
